backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.13.3
python-multipart==0.0.9
boto3==1.35.36
prometheus-fastapi-instrumentator==6.1.0
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0